+++ lightning/src/chain/chaininterface.rs
@@ -119,7 +119,7 @@ pub trait FeeEstimator: Sync + Send {
 	/// This translates to:
 	///  * satoshis-per-byte * 250
 	///  * ceil(satoshis-per-kbyte / 4)
-	fn get_est_sat_per_1000_weight(&self, confirmation_target: ConfirmationTarget) -> u64;
+	fn get_est_sat_per_1000_weight(&self, confirmation_target: ConfirmationTarget) -> u32;
 }
 
 /// Minimum relay fee as required by bitcoin network mempool policy.
